fix: 사이드바 컴포넌트 통합 및 파일명 변경
